extract/requirements.txt
fastapi
uvicorn[standard]
confluent-kafka
python-dotenv
pydantic
